chore: create web-components package
