+++ packages/wallet-connection/index.js
@@ -1,1 +1,2 @@
-export { AgoricWalletConnection } from './src/AgoricWalletConnection.js';
+export { AgoricWalletConnection } from '@agoric/web-components';
+console.warn('This package is deprecated. Use @agoric/web-components instead of @agoric/wallet-connection.');
