+++ packages/web-components/agoric-wallet-connection.js
@@ -1,0 +1,7 @@
+import { AgoricWalletConnection } from './src/AgoricWalletConnection.js';
+import './src/agoric-iframe-messenger.js';
+
+window.customElements.define(
+  'agoric-wallet-connection',
+  AgoricWalletConnection,
+);
