+++ packages/web-components/demo/install-ses-lockdown.js
