+++ packages/web-components/index.js
@@ -1,0 +1,1 @@
+export { AgoricWalletConnection } from './src/AgoricWalletConnection.js';
