+++ packages/web-components/src/AgoricWalletConnection.js
